requirements.txt
# Deployment
gunicorn==20.0.4

# App server
Flask==1.1.2
pymongo==3.10.1
flask-httpauth==4.1.0
flask-limiter==1.3.1
python-dotenv==0.13.0